pyproject.toml
[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["wsws_bulletin*"]
exclude = ["output*"]

[project]
name = "wsws-bulletin"
version = "0.1.0"
description = "An automatic daily bulletin based on the World Socialist Web Site"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "WSWS Bulletin"}
]
dependencies = [
    "requests>=2.31.0,<3.0.0",
    "requests-cache>=1.2.0,<2.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "lxml>=4.9.0,<5.0.0",
    "click>=8.1.0,<9.0.0",
    "openai>=1.12.0,<2.0.0",
    "anthropic>=0.72.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "coqui-tts>=0.24.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "black>=23.0.0,<25.0.0",
    "ruff>=0.1.0,<1.0.0",
]
tts-coqui = [
    "coqui-tts>=0.24.0,<1.0.0",
]
tts-all = [
    "coqui-tts>=0.24.0,<1.0.0",
]

[project.scripts]
wsws-bulletin = "wsws_bulletin.__main__:main"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
